+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - programs(store): Added `cumulative_inv_cost_factor` as a new global metric for GT, along with an instruction to update it.
+- programs(store): Added `mint_gt_reward` instruction.
 
 ## [0.7.1] - 2025-08-15